Deprecate torch.triangular_solve
There is a use of `at::triangular_solve_out` in the file
`torch/csrc/jit/tensorexpr/external_functions.cpp` that I have not dared
to move to `at::linalg_triangular_solve_out`.

**BC-breaking note:**

This PR deprecates the `torch.triangular_solve` function in favor of
`torch.linalg.triangular_solve`. An upgarde guide is added to the
documentation for `torch.triangular_solve`.

Note that it DOES NOT remove `torch.triangular_solve`, but
`torch.triangular_solve` will be removed in a future PyTorch release.

ghstack-source-id: c5afcdba235a6b081e5f8b09b384eb8fb048375f
Pull Request resolved: https://github.com/pytorch/pytorch/pull/63570