Rename int to long, add more C++ types.
BC-breaking change: intT is now longT (which aligns it more accurately with how
the types are referred to in C++).  The benefit for this is we can idiomatically
express all C++ dtypes (with intT now mapping to int32_t).  These types are needed
for ufunc codegen in a latter patch.

Signed-off-by: Edward Z. Yang <ezyang@fb.com>

ghstack-source-id: 5619303a12b016bdb82e913f7c48ee53aad81e5a
Pull Request resolved: https://github.com/pytorch/pytorch/pull/66108