[sparsity] Fix GPU training for sparsity
The GPU training was not supported in the sparsifier.
The reason was that when the sparsifier was created the masks would default to the CPU.
Attaching a GPU model to the sparsifier would throw an error.
The solution is to create the masks on the same device as the weight.

ghstack-source-id: 80c00df3740ad833c5815fa366475c4dad1f6208
Pull Request resolved: https://github.com/pytorch/pytorch/pull/66412