[PyTorch] Capture function args from schema by reference
Pull Request resolved: https://github.com/pytorch/pytorch/pull/65951

Profiling shows that we do a bunch of heap allocations to copy Argument structs in append_operator. Capturing by reference here should be safe as long as the schema objects args is from outlive the operator function.

IMPORTANT: Reviewers (or automated tests if we're lucky) need to
confirm that the above is true or we're going to have fun
use-after-free bugs.
ghstack-source-id: 140375637

Differential Revision: [D31319988](https://our.internmc.facebook.com/intern/diff/D31319988/)