[TensorExpr] Add lowering for aten::max (reduction).
Is it a hack?
Yes and no. Yes, because I had to disable lowering for pointwise max.
No, because eventually we would need to add max reduction lowering
anyway. A proper implementation requires #65843 - that would allow us to
distinguish pointwise case from the reduction case and have separate
lowerings for them. In the long run we should land #65843 and then a
cleaned up version of this PR.

ghstack-source-id: f04b5b75ae68b6e78e6c7c3798848e5399e444b2
Pull Request resolved: https://github.com/pytorch/pytorch/pull/66519